+++ lib/dradis/plugins/html_export/exporter.rb
@@ -28,9 +28,7 @@ module Dradis
         end
 
         def nodes
-          # FIXME: This is an ugly piece of code and the list of nodes should
-          # come from the ContentService.
-          @nodes ||= issues.map(&:evidence).flatten.map(&:node).uniq
+          @nodes ||= content_service.all_evidence.map(&:node).uniq
         end
 
         def notes