+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 [v#.#.#] ([month] [YYYY])
-  - Bugs fixed:
-    - Only include Published Evidence in exports when the export scope is set to Published
+  - Only include Published Evidence in exports when the export scope is set to Published
 
 v5.3.0 (August 2026)
   - No changes