+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-[v#.#.#] ([month] [YYYY])
+v5.4.0 (September 2026)
   - Only include Published Evidence in exports when the export scope is set to Published
 
 v5.3.0 (August 2026)